content-security-policy/script-src/script-src-strict_dynamic_parser_inserted.html

<!DOCTYPE HTML>
<html>

<head>
    <title>Parser-inserted scripts without a correct nonce are not allowed with `strict-dynamic` in the script-src directive.</title>
    <script src='/resources/testharness.js' nonce='dummy'></script>
    <script src='/resources/testharnessreport.js' nonce='dummy'></script>

    <!-- CSP served: script-src 'strict-dynamic' 'nonce-dummy' -->
</head>

<body>
    <h1>Parser-inserted scripts without a correct nonce are not allowed with `strict-dynamic` in the script-src directive.</h1>
    <div id='log'></div>

    <script nonce='dummy'>
        async_test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('documentWrite').getAttribute('executed'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            document.write('<scr' + 'ipt id="documentWrite" src="simpleSourcedScript.js"></scr' + 'ipt>');
        }, 'Parser-inserted script via `document.write` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('documentWriteln').getAttribute('executed'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            document.writeln('<scr' + 'ipt id="documentWriteln" src="simpleSourcedScript.js"></scr' + 'ipt>');
        }, 'Parser-inserted script via `document.writeln` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('documentWrite-deferred').getAttribute('executed'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            document.write('<scr' + 'ipt defer id="documentWrite-deferred" src="simpleSourcedScript.js"></scr' + 'ipt>');
        }, 'Parser-inserted deferred script via `document.write` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('documentWriteln-deferred').getAttribute('executed'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            document.writeln('<scr' + 'ipt defer id="documentWriteln-deferred" src="simpleSourcedScript.js"></scr' + 'ipt>');
        }, 'Parser-inserted deferred script via `document.writeln` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('documentWrite-async').getAttribute('executed'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            document.write('<scr' + 'ipt async id="documentWrite-async" src="simpleSourcedScript.js"></scr' + 'ipt>');
        }, 'Parser-inserted async script via `document.write` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('documentWriteln-async').getAttribute('executed'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            document.writeln('<scr' + 'ipt async id="documentWriteln-async" src="simpleSourcedScript.js"></scr' + 'ipt>');
        }, 'Parser-inserted async script via `document.writeln` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('documentWrite-deferred-async').getAttribute('executed'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            document.write('<scr' + 'ipt defer async id="documentWrite-deferred-async" src="simpleSourcedScript.js"></scr' + 'ipt>');
        }, 'Parser-inserted deferred async script via `document.write` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('documentWriteln-deferred-async'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            document.writeln('<scr' + 'ipt defer async id="documentWriteln-deferred-async " src="simpleSourcedScript.js"></scr' + 'ipt>');
        }, 'Parser-inserted deferred async script via `document.writeln` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        var innerHTMLScriptRan = false;
        test(function(t) {
            var handler = function(e) {
                assert_false(innerHTMLScriptRan);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);
            var e = document.createElement('div');
            e.innerHTML = "<img src='/nonexisting.jpg' onerror='innerHTMLScriptRan = true;' style='display:none'>";
            document.body.appendChild(e);
        }, 'Script injected via `innerHTML` is not allowed with `strict-dynamic`.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        var insertAdjacentHTMLScriptRan = false;
        test(function(t) {
            var handler = function(e) {
                assert_false(insertAdjacentHTMLScriptRan);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);
            var e = document.createElement('div');
            e.insertAdjacentHTML('afterbegin', "<img src='/nonexisting.jpg' onerror='insertAdjacentHTMLScriptRan = true;' style='display:none'>");
            document.body.appendChild(e);
        }, 'Script injected via `insertAdjacentHTML` is not allowed with `strict-dynamic`.');
    </script>
</body>

</html>
